Move the internal stuff up one level.
Because a higher-level internal directory is introduced
in #194.
