+++ rules/aip0131/method_rules.go
@@ -17,8 +17,8 @@ package aip0131
 import (
 	"fmt"
 
+	"github.com/googleapis/api-linter/internal/utils"
 	"github.com/googleapis/api-linter/lint"
-	"github.com/googleapis/api-linter/rules/internal/utils"
 	"github.com/jhump/protoreflect/desc"
 )
