--- rules/internal/utils/doc.go
@@ -1,2 +1,0 @@
-// Package utils contains helper functions for writing rules.
-package utils
